# backend-node/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/users/access": {
      "get": {
        "operationId": "AdminController_listAccess",
        "parameters": [
          {
            "name": "filter",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{userId}/roles/{roleName}": {
      "post": {
        "operationId": "AdminController_assignRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_removeRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/applications": {
      "post": {
        "operationId": "ApplicationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      },
      "get": {
        "operationId": "ApplicationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/applications/{id}": {
      "get": {
        "operationId": "ApplicationsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      },
      "put": {
        "operationId": "ApplicationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      },
      "delete": {
        "operationId": "ApplicationsController_deleteApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{id}/submit": {
      "post": {
        "operationId": "ApplicationsController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{id}/status": {
      "post": {
        "operationId": "ApplicationsController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{id}/consent": {
      "post": {
        "operationId": "ApplicationsController_recordConsent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordConsentDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{id}/history": {
      "get": {
        "operationId": "ApplicationsController_history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{id}/notes": {
      "get": {
        "operationId": "ApplicationsController_listNotes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      },
      "post": {
        "operationId": "ApplicationsController_createNote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{id}/documents/presign-upload": {
      "post": {
        "operationId": "ApplicationsController_presignUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{id}/documents/confirm": {
      "post": {
        "operationId": "ApplicationsController_confirmUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{id}/documents": {
      "get": {
        "operationId": "ApplicationsController_listDocuments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/loans/{id}": {
      "get": {
        "operationId": "LoansController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "loans"
        ]
      }
    },
    "/api/loans/{id}/disburse": {
      "post": {
        "operationId": "LoansController_disburse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisburseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "loans"
        ]
      }
    },
    "/api/loans/{id}/repayments": {
      "post": {
        "operationId": "LoansController_recordRepayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordRepaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "loans"
        ]
      }
    },
...
    "/api/document-requirements": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequirementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      }
    },
    "/api/applications/{appId}/documents/{docId}/verify": {
      "post": {
        "operationId": "DocumentsController_verify",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyDocumentDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      }
    },
    "/api/applications/{appId}/documents/{docId}": {
      "delete": {
        "operationId": "DocumentsController_deleteDocument",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      }
    },
    "/api/applications/{appId}/documents/{docId}/url": {
      "get": {
        "operationId": "DocumentsController_getSignedUrl",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "PRDF LMS API",
    "description": "Loan origination and servicing API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateApplicationDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "requestedAmount": {
            "type": "number"
          },
          "termMonths": {
            "type": "number"
          },
          "purpose": {
            "type": "string"
          },
          "businessName": {
            "type": "string"
          },
          "registrationNo": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "assignedToUserId": {
            "type": "string"
          },
          "monthlyRevenue": {
            "type": "number"
          },
          "yearsInOperation": {
            "type": "number"
          },
          "numberOfEmployees": {
            "type": "number"
          },
          "bankName": {
            "type": "string"
          },
          "currentStep": {
            "type": "number"
          },
          "draftState": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "UpdateApplicationDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "requestedAmount": {
            "type": "number"
          },
          "termMonths": {
            "type": "number"
          },
          "purpose": {
            "type": "string"
          },
          "businessName": {
            "type": "string"
          },
          "registrationNo": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "assignedToUserId": {
            "type": "string"
          },
          "monthlyRevenue": {
            "type": "number"
          },
          "yearsInOperation": {
            "type": "number"
          },
          "numberOfEmployees": {
            "type": "number"
          },
          "bankName": {
            "type": "string"
          },
          "currentStep": {
            "type": "number"
          },
          "draftState": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "SubmitApplicationDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          }
        }
      },
      "ChangeStatusDto": {
        "type": "object",
        "properties": {
          "toStatus": {
            "type": "string",
            "enum": [
              "Draft",
              "Submitted",
              "UnderReview",
              "InfoRequested",
              "Approved",
              "Rejected",
              "Disbursed",
              "InRepayment",
              "Closed"
            ]
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "toStatus"
        ]
      },
      "RecordConsentDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "description": "Consent copy/version identifier, e.g. \"2026-07-01\"."
          },
          "items": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "version",
          "items"
        ]
      },
      "CreateNoteDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        },
        "required": [
          "body"
        ]
      },
      "PresignUploadDto": {
        "type": "object",
        "properties": {
          "docType": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          }
        },
        "required": [
          "docType",
          "fileName"
        ]
      },
      "ConfirmUploadDto": {
        "type": "object",
        "properties": {
          "docType": {
            "type": "string"
          },
          "storagePath": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "docType",
          "storagePath"
        ]
      },
      "DisburseDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "reference": {
            "type": "string"
          }
        },
        "required": [
          "amount"
        ]
      },
      "RecordRepaymentDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "paidAt": {
            "type": "string",
            "description": "ISO date-time; defaults to now if omitted."
          },
          "paymentReference": {
            "type": "string"
          }
        },
        "required": [
          "amount"
        ]
      },
...
      "CreateRequirementDto": {
        "type": "object",
        "properties": {
          "loanProductId": {
            "type": "string"
          },
          "requiredAtStatus": {
            "type": "string"
          },
          "docType": {
            "type": "string"
          },
          "isRequired": {
            "type": "boolean"
          }
        },
        "required": [
          "requiredAtStatus",
          "docType",
          "isRequired"
        ]
      },
      "VerifyDocumentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "Verified",
              "Rejected"
            ]
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}