# backend-node/src/applications/dto/submit-application.dto.ts
import { ApiPropertyOptional } from '@nestjs/swagger';
import { IsOptional, IsString } from 'class-validator';

export class SubmitApplicationDto {
  @ApiPropertyOptional()
  @IsOptional()
  @IsString()
  note?: string;
}